Upgrade Faraday to 2.6.0
Upgrades Faraday to 2.6.0 and replaces faraday_middleware with
individual gems that have been extracted from the deprecated gem.

Custom middleware has been updated to subclass Faraday::Middleware
instead of Faraday::Response::Middleware.

Faraday 2 returns a nil body for DELETE requests so specs have been
updated to match since DELETE requests do not typically have a response
body. This is a breaking change so it might be better to change the
response body to parse as an empty hash instead, which matches the
previous behaviour.